Move types to relevant slice folders.
