+++ src/app/slices/maps/types.ts
@@ -1,0 +1,5 @@
+export type MapResult = {
+  lat: string;
+  lng: string;
+  name: string;
+};
